backend/requirements.txt
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
ecdsa==0.19.1
Flask==3.1.1
greenlet==3.2.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
playwright==1.52.0
pyee==13.0.0
python-dotenv==1.1.0
python-http-client==3.3.7
requests==2.32.3
sendgrid==6.12.2
six==1.17.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==2.3.8
